arreglado el error del tipo de la matriz A

+++ src/main.h
@@ -68,4 +68,4 @@ float getNextFloatFromInputFile(ifstream & inputFile){
 
 
 void handleError(int);
-vector< vector< int > > crearMatrizA (int cantAngulos, int cantRadios, int comienzoPared, int finalPared);
+vector< vector< float > > crearMatrizA (int cantAngulos, int cantRadios, int comienzoPared, int finalPared);
